feat: 주식 기능 구현
- 매도, 매수시 userId sharedpreference에서 받아오기

S12P11D107-333 #done 동작 확인 완료

+++ frontend/Ttaenggrang/app/src/main/java/com/ladysparks/ttaenggrang/data/remote/StockService.kt
@@ -14,9 +14,9 @@ interface StockService {
     // 주식 매도
     @POST("stocks/{stockId}/sell")  // ✅ 경로 확인
     suspend fun sellStock(
-        @Path("stockId") stockId: Long,
+        @Path("stockId") stockId: Int,
         @Query("share_count") shareCount: Int,
-        @Query("studentId") studentId: Long
+        @Query("studentId") studentId: Int
     ): Response<StockTransactionResponse>
 
     // 주식 매수
